requirements.txt
pyserial>=3.5
matplotlib>=3.5.0
seaborn>=0.11.2
numpy>=1.21.0
